+++ sources/clear.rb
@@ -23,7 +23,8 @@ def clear (db)
         confirm_clear(db, name[0])
       end
     else
-      abort("error while trying to get content of file ( -f option )")
+      puts "error while trying to get content of file ( -f option )"
+      exit 5
     end
   end
 end