+++ sources/utils_db.rb
@@ -1,0 +1,5 @@
+def db_error_exit(message, error)
+  puts message
+  puts "message is : '" + error.message + "'"
+  exit 2
+end